# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/admin/clinic-settings": {
      "get": {
        "operationId": "ClinicSettingsController_getSettings",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClinicSettingsDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Read the clinic-wide settings and working hours",
        "tags": [
          "admin/clinic-settings"
        ]
      },
      "put": {
        "operationId": "ClinicSettingsController_updateSettings",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateClinicSettingsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClinicSettingsDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update the clinic-wide settings and working hours",
        "tags": [
          "admin/clinic-settings"
        ]
      }
    },
...
    "/api/v1/coupons/validate": {
      "post": {
        "description": "Read-only — never increments used_count.",
        "operationId": "CouponsController_validate",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ValidateCouponDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CouponPreviewResponseDto"
                }
              }
            }
          }
        },
        "summary": "Preview a coupon against a service price",
        "tags": [
          "coupons"
        ]
      }
    },
...
  },
  "info": {
    "title": "Physio Clinic API",
    "description": "Booking, patient records and clinic administration for the physiotherapy clinic. This specification is the contract the website and CRM generate their types from.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "CRM access token from POST /auth/google or /auth/login."
      }
    },
    "schemas": {
...
      "WorkingIntervalDto": {
        "type": "object",
        "properties": {
          "start": {
            "type": "string",
            "example": "09:00"
          },
          "end": {
            "type": "string",
            "example": "13:00"
          }
        },
        "required": [
          "start",
          "end"
        ]
      },
      "WeeklyWorkingHoursDto": {
        "type": "object",
        "properties": {
          "mon": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WorkingIntervalDto"
            }
          },
          "tue": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WorkingIntervalDto"
            }
          },
          "wed": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WorkingIntervalDto"
            }
          },
          "thu": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WorkingIntervalDto"
            }
          },
          "fri": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WorkingIntervalDto"
            }
          },
          "sat": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WorkingIntervalDto"
            }
          },
          "sun": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WorkingIntervalDto"
            }
          }
        },
        "required": [
          "mon",
          "tue",
          "wed",
          "thu",
          "fri",
          "sat",
          "sun"
        ]
      },
      "ClinicSettingsDto": {
        "type": "object",
        "properties": {
          "timezone": {
            "type": "string",
            "example": "Asia/Kolkata"
          },
          "workingHours": {
            "$ref": "#/components/schemas/WeeklyWorkingHoursDto"
          },
          "slotBufferMinutes": {
            "type": "number",
            "example": 0
          }
        },
        "required": [
          "timezone",
          "workingHours",
          "slotBufferMinutes"
        ]
      },
      "UpdateClinicSettingsDto": {
        "type": "object",
        "properties": {
          "workingHours": {
            "$ref": "#/components/schemas/WeeklyWorkingHoursDto"
          },
          "slotBufferMinutes": {
            "type": "number",
            "example": 0
          },
          "timezone": {
            "type": "string",
            "example": "Asia/Kolkata"
          }
        }
      },
...
      "ValidateCouponDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "example": "WELCOME10"
          },
          "serviceId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "code",
          "serviceId"
        ]
      },
      "CouponPreviewResponseDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "example": "WELCOME10"
          },
          "valueType": {
            "type": "string",
            "enum": [
              "percent",
              "fixed"
            ]
          },
          "value": {
            "type": "string",
            "example": "10.00"
          },
          "originalPrice": {
            "type": "string",
            "example": "1000.00"
          },
          "discountAmount": {
            "type": "string",
            "example": "100.00"
          },
          "finalPrice": {
            "type": "string",
            "example": "900.00"
          }
        },
        "required": [
          "code",
          "valueType",
          "value",
          "originalPrice",
          "discountAmount",
          "finalPrice"
        ]
      },
...
    }
  }
}